# public/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ChatGPT 聊天 API",
        "description": "這是我的 Laravel + Swagger API 文件示範",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "使用者註冊",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "account",
                                    "password"
                                ],
                                "properties": {
                                    "account": {
                                        "type": "string",
                                        "example": "joe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "0111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "註冊成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User registered successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "使用者登入",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "account",
                                    "password"
                                ],
                                "properties": {
                                    "account": {
                                        "type": "string",
                                        "example": "joe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "0111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "登入成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "帳密錯誤"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "使用者登出",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "登出成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user-list": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "查看所有使用者帳號",
                "operationId": "b6ef6f5fe7c8430d6493a12ce5dae2c1",
                "responses": {
                    "200": {
                        "description": "成功回傳帳號清單",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "account": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/messages": {
            "get": {
                "tags": [
                    "Message"
                ],
                "summary": "取得使用者的歷史訊息",
                "operationId": "9d5effa2cb648df85be0ea9741ffb457",
                "responses": {
                    "200": {
                        "description": "成功回傳訊息列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Message"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/message": {
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "送出訊息給 ChatGPT 並儲存紀錄",
                "operationId": "ee15771b1dadd0efb04ebd263be4f168",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "你好，今天幾號？"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "GPT 回應的訊息",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reply": {
                                            "type": "string",
                                            "example": "今天是 4 月 9 日。"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未授權"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/test-post": {
            "post": {
                "tags": [
                    "Swagger"
                ],
                "summary": "測試 POST",
                "description": "測試 Swagger 的 POST API",
                "operationId": "cdc72eb04ce4f53e4c1f03bc7bf7df05",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Joey"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joey@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功回應",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "收到資料！"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Message": {
                "title": "訊息",
                "required": [
                    "user_id",
                    "role",
                    "content"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "role": {
                        "type": "string",
                        "example": "user"
                    },
                    "content": {
                        "type": "string",
                        "example": "你好"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-09T15:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-09T15:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Message",
            "description": "Message"
        },
        {
            "name": "Swagger",
            "description": "Swagger"
        }
    ]
}